+++ src/app/services/account.service.ts
@@ -3,7 +3,7 @@ import { SpyglassService } from './spyglass.service';
 import { RpcService } from '@app/services/rpc.service';
 import { AppStateService, AppStore } from '@app/services/app-state.service';
 import { AccountOverview } from '@app/types/AccountOverview';
-import {ReceivableTx} from "@app/types/ReceivableTx";
+import { ReceivableTx } from '@app/types/ReceivableTx';
 
 @Injectable({
     providedIn: 'root',