+++ CHANGELOG.md
@@ -4,13 +4,13 @@
 
 ### Added
 
--  Added toggle to enable / disable automatic receiving.
+-  Added toggle to enable / disable automatic receiving, defaulting to Enabled.
 
 ## v4.9.0 (February 5, 2024)
 
 ### Added
 
--  Added ability to select which transactions to receive. 
+-  Added ability to select which transactions to receive from within Receive transaction overlay.  
 
 ## v4.8.0 (January 7, 2024)